feat: AuthJS and Bungie login implementation
