+++ src/error.html
@@ -1,24 +1,11 @@
-<!doctype html>
-<html lang="en">
-	<head>
-		<meta charset="UTF-8" />
-		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<title>Error</title>
-		<link
-			href="https://cdn.jsdelivr.net/npm/tailwindcss@2.2.19/dist/tailwind.min.css"
-			rel="stylesheet"
-		/>
-		<script src="https://cdn.jsdelivr.net/npm/daisyui@2.51.5/dist/full.js"></script>
-	</head>
-	<body class="flex min-h-screen items-center justify-center bg-gray-100">
-		<div class="card w-96 bg-white shadow-xl">
-			<div class="card-body items-center text-center">
-				<h2 class="card-title text-4xl text-red-500">Error 404</h2>
-				<p class="text-gray-600">Oops! The page you're looking for doesn't exist.</p>
-				<div class="card-actions mt-4 justify-center">
-					<a href="/" class="btn btn-primary">Go Home</a>
-				</div>
+<div class="flex min-h-screen items-center justify-center bg-gray-100">
+	<div class="card w-96 bg-white shadow-xl">
+		<div class="card-body items-center text-center">
+			<h2 class="card-title text-4xl text-red-500">Error 404</h2>
+			<p class="text-gray-600">Oops! The page you're looking for doesn't exist.</p>
+			<div class="card-actions mt-4 justify-center">
+				<a href="/" class="btn btn-primary">Go Home</a>
 			</div>
 		</div>
-	</body>
-</html>
+	</div>
+</div>
