+++ src/hooks.server.ts
@@ -1,4 +1,7 @@
-import type { Handle } from '@sveltejs/kit';
 import { i18n } from '$lib/i18n';
+import type { Handle } from '@sveltejs/kit';
+import { sequence } from '@sveltejs/kit/hooks';
+import { handle as handleAuthJS } from './auth';
 const handleParaglide: Handle = i18n.handle();
-export const handle: Handle = handleParaglide;
+// Handle sequence of AuthJS and Paraglide
+export const handle: Handle = sequence(handleAuthJS, handleParaglide);
