+++ src/routes/+layout.server.ts
@@ -1,0 +1,9 @@
+import type { LayoutServerLoad } from './$types';
+
+export const load: LayoutServerLoad = async (event) => {
+	const session = await event.locals.auth();
+
+	return {
+		session
+	};
+};
