+++ src/routes/+page.svelte
@@ -1,30 +1,11 @@
 <script lang="ts">
 	import * as m from '$lib/paraglide/messages';
-	import { getPlayerProfile } from '$lib/services/bungieApi';
-
-	let playerProfile: any = null;
-
-	const fetchPlayerProfile = async () => {
-		try {
-			// Replace with actual membershipType, membershipId, and destinyProfileId
-			const membershipType = 1; // Example: Xbox
-			const membershipId = '123456789';
-			const destinyProfileId = '987654321';
-
-			playerProfile = await getPlayerProfile(membershipType, membershipId, destinyProfileId);
-		} catch (error) {
-			console.error('Failed to fetch player profile:', error);
-		}
-	};
-
-	fetchPlayerProfile();
+	import { signIn } from '@auth/sveltekit/client';
 </script>
 
-<h1>{m.hello_world({ name: 'user' })}</h1>
-<p>Visit <a href="https://svelte.dev/docs/kit">svelte.dev/docs/kit</a> to read the documentation</p>
-
-{#if playerProfile}
-	<pre>{JSON.stringify(playerProfile, null, 2)}</pre>
-{:else}
-	<p>{m.loading_player()}</p>
-{/if}
+<div class="flex flex-col items-center justify-center min-h-screen bg-base-200">
+	<!-- <img src="/img/logo.svg" alt="Company Logo" class="w-32 h-32 mb-6" /> -->
+	<button class="btn btn-primary" on:click={() => signIn()}>
+		{m.sign_in()}
+	</button>
+</div>
