+++ vite.config.ts
@@ -1,6 +1,6 @@
 import { paraglide } from '@inlang/paraglide-sveltekit/vite';
-import tailwindcss from '@tailwindcss/vite';
 import { sveltekit } from '@sveltejs/kit/vite';
+import tailwindcss from '@tailwindcss/vite';
 import { defineConfig } from 'vite';
 
 export default defineConfig({
